fix: fix reaction alignment and theming.
