+++ packages/stream_chat_flutter/lib/src/misc/reaction_icon.dart
@@ -1,5 +1,14 @@
 import 'package:flutter/material.dart';
 
+/// {@template reactionIconBuilder}
+/// Signature for a function that builds a reaction icon.
+/// {@endtemplate}
+typedef ReactionIconBuilder = Widget Function(
+  BuildContext context,
+  bool isHighlighted,
+  double iconSize,
+);
+
 /// {@template streamReactionIcon}
 /// Reaction icon data
 /// {@endtemplate}
@@ -13,10 +22,6 @@ class StreamReactionIcon {
   /// Type of reaction
   final String type;
 
-  /// Asset to display for reaction
-  final Widget Function(
-    BuildContext,
-    bool highlighted,
-    double size,
-  ) builder;
+  /// {@macro reactionIconBuilder}
+  final ReactionIconBuilder builder;
 }
